Apply page and some minor changes
